Update spellcard.py
I revised my spell card file so it would work with Thomas' unit card code. I added a status effect function to make changes to the stats easier. The sight card to reveal the opponent's deck and the scissor card to remove a card from the opponent's deck isn't implemented yet, but it feels like it falls under the game logic?